+++ lib/Imagine/Gmagick/Imagine.php
@@ -14,7 +14,7 @@ namespace Imagine\Gmagick;
 use Imagine\Exception\InvalidArgumentException;
 use Imagine\Exception\NotSupportedException;
 use Imagine\Exception\RuntimeException;
-use Imagine\File\Loader;
+use Imagine\Factory\ClassFactoryInterface;
 use Imagine\File\LoaderInterface;
 use Imagine\Image\AbstractImagine;
 use Imagine\Image\BoxInterface;
@@ -45,13 +45,13 @@ class Imagine extends AbstractImagine
      */
     public function open($path)
     {
-        $loader = $path instanceof LoaderInterface ? $path : new Loader($path);
+        $loader = $path instanceof LoaderInterface ? $path : $this->getClassFactory()->createFileLoader($path);
         $path = $loader->getPath();
 
         try {
             if ($loader->isLocalFile()) {
                 $gmagick = new \Gmagick($path);
-                $image = new Image($gmagick, $this->createPalette($gmagick), $this->getMetadataReader()->readFile($loader));
+                $image = $this->getClassFactory()->createImage(ClassFactoryInterface::HANDLE_GMAGICK, $gmagick, $this->createPalette($gmagick), $this->getMetadataReader()->readFile($loader));
             } else {
                 $image = $this->doLoad($loader->getData(), $this->getMetadataReader()->readFile($loader));
             }
@@ -94,7 +94,7 @@ class Imagine extends AbstractImagine
             $gmagick->setimagecolorspace(\Gmagick::COLORSPACE_TRANSPARENT);
             $gmagick->setimagebackgroundcolor($pixel);
 
-            $image = new Image($gmagick, $palette, new MetadataBag());
+            $image = $this->getClassFactory()->createImage(ClassFactoryInterface::HANDLE_GMAGICK, $gmagick, $palette, new MetadataBag());
 
             if ($switchPalette) {
                 $image->usePalette($switchPalette);
@@ -137,10 +137,7 @@ class Imagine extends AbstractImagine
      */
     public function font($file, $size, ColorInterface $color)
     {
-        $gmagick = new \Gmagick();
-        $gmagick->newimage(1, 1, 'transparent');
-
-        return new Font($gmagick, $file, $size, $color);
+        return $this->getClassFactory()->createFont(ClassFactoryInterface::HANDLE_GMAGICK, $file, $size, $color);
     }
 
     private function createPalette(\Gmagick $gmagick)
@@ -167,6 +164,6 @@ class Imagine extends AbstractImagine
             throw new RuntimeException('Could not load image from string', $e->getCode(), $e);
         }
 
-        return new Image($gmagick, $this->createPalette($gmagick), $metadata);
+        return $this->getClassFactory()->createImage(ClassFactoryInterface::HANDLE_GMAGICK, $gmagick, $this->createPalette($gmagick), $metadata);
     }
 }