+++ lib/Imagine/Image/AbstractImage.php
@@ -12,14 +12,22 @@
 namespace Imagine\Image;
 
 use Imagine\Exception\InvalidArgumentException;
+use Imagine\Factory\ClassFactory;
+use Imagine\Factory\ClassFactoryAwareInterface;
+use Imagine\Factory\ClassFactoryInterface;
 use Imagine\Image\Metadata\MetadataBag;
 
-abstract class AbstractImage implements ImageInterface
+abstract class AbstractImage implements ImageInterface, ClassFactoryAwareInterface
 {
     /**
      * @var MetadataBag
      */
     protected $metadata;
+
+    /**
+     * @var \Imagine\Factory\ClassFactoryInterface|null
+     */
+    private $classFactory;
 
     /**
      * {@inheritdoc}
@@ -115,5 +123,31 @@ abstract class AbstractImage implements ImageInterface
         if ($this->metadata !== null) {
             $this->metadata = clone $this->metadata;
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Factory\ClassFactoryAwareInterface::getClassFactory()
+     */
+    public function getClassFactory()
+    {
+        if ($this->classFactory === null) {
+            $this->classFactory = new ClassFactory();
+        }
+
+        return $this->classFactory;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Factory\ClassFactoryAwareInterface::setClassFactory()
+     */
+    public function setClassFactory(ClassFactoryInterface $classFactory)
+    {
+        $this->classFactory = $classFactory;
+
+        return $this;
     }
 }