+++ lib/Imagine/Image/ImagineInterface.php
@@ -13,12 +13,13 @@ namespace Imagine\Image;
 
 use Imagine\Exception\InvalidArgumentException;
 use Imagine\Exception\RuntimeException;
+use Imagine\Factory\ClassFactoryAwareInterface;
 use Imagine\Image\Palette\Color\ColorInterface;
 
 /**
  * The imagine interface.
  */
-interface ImagineInterface
+interface ImagineInterface extends ClassFactoryAwareInterface
 {
     const VERSION = '0.7-dev';