+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   * Fix warning about deprecated clone method in copy method of Imagick images (@mlocati)
   * Fix copy methods of Images (the original image and its new copy are now fully detached) (@mlocati)
   * It's now possible to use `clone $image` as an alternative to `$image->copy()` (@mlocati)
+  * Add support to custom classes for BoxInterface, MetadataReaderInterface, FontInterface, LoaderInterface, LayersInterface, ImageInterface (@mlocati)
 
 ### 0.7.1 (2017-05-16)
   * Remove Symfony PHPUnit bridge as dependency (@craue)